more methods into class dirs
