+++ src/image/toLatex.js
@@ -1,0 +1,12 @@
+//
+const toLatex = function(image) {
+  let alt = image.file.replace(/^(file|image):/i, '');
+  alt = alt.replace(/\.(jpg|jpeg|png|gif|svg)/i, '');
+  var out = '\\begin{figure}';
+  out += '\n\\includegraphics[width=\\linewidth]{' + image.thumb + '}';
+  out += '\n\\caption{' + alt + '}';
+  out += '\n%\\label{fig:myimage1}';
+  out += '\n\\end{figure}';
+  return out;
+};
+module.exports = toLatex;
